Fix format specifier for 64-bit index in error messages
Commit e2ab21f2 ("Implement GL_EXT_shader_64bit_indexing") made the
index int64_t but left the "index out of range" messages using '%d'.

32-bit ARM and PowerPC align 64-bit variadic arguments, so '%d' reads
the skipped padding (a stack slot on ARM, a register on PowerPC) and
prints garbage. This breaks the runtests single-thread vs. multithread
output comparison.

Fixes #4180

+++ glslang/MachineIndependent/ParseHelper.cpp
@@ -649,15 +649,15 @@ TIntermTyped* TParseContext::handleBracketDereference(const TSourceLoc& loc, TIn
             base->getWritableType().setImplicitlySized(true);
             if (base->getQualifier().builtIn == EbvClipDistance &&
                 indexValue >= resources.maxClipDistances) {
-                error(loc, "gl_ClipDistance", "[", "array index out of range '%d'", indexValue);
+                error(loc, "gl_ClipDistance", "[", "array index out of range '%lld'", (long long)indexValue);
             }
             else if (base->getQualifier().builtIn == EbvCullDistance &&
                 indexValue >= resources.maxCullDistances) {
-                error(loc, "gl_CullDistance", "[", "array index out of range '%d'", indexValue);
+                error(loc, "gl_CullDistance", "[", "array index out of range '%lld'", (long long)indexValue);
             }
             else if (base->getQualifier().builtIn == EbvSampleMask &&
                 indexValue >= (resources.maxSamples + 31) / 32) {
-                error(loc, "gl_SampleMask", "[", "array index out of range '%d'", indexValue);
+                error(loc, "gl_SampleMask", "[", "array index out of range '%lld'", (long long)indexValue);
             }
             // For 2D per-view builtin arrays, update the inner dimension size in parent type
             if (base->getQualifier().isPerView() && base->getQualifier().builtIn != EbvNone) {
